requirements.txt
boto3==1.16.37
botocore==1.19.37
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
coloredlogs==15.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
Flask==1.1.2
grpcio==1.34.0
humanfriendly==9.1
idna==2.10
imageio==2.9.0
influxdb==5.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
msgpack==1.0.1
name-server==2.2.1
networkx==2.5
nori==0.9.9
nori2==1.11.4
numpy==1.19.4
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
Pillow==8.0.1
prettytable==2.0.0
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-statemachine==0.8.0
pytz==2020.4
PyWavelets==1.1.1
redis==3.5.3
refile==5.3.2
requests==2.25.0
s3transfer==0.3.3
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
smart-open==4.0.1
tabulate==0.8.7
termcolor==1.1.0
texttable==1.6.3
tifffile==2020.9.3
torch==1.7.1
torchvision==0.8.2
tqdm==4.54.1
typing-extensions==3.7.4.3
ujson==4.0.1
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
